use size_t

+++ src/blocklist.t.c
@@ -6,8 +6,9 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <assert.h>
+#include <stddef.h>
 
-void* Demo_Malloc(unsigned long nBytes){
+void* Demo_Malloc(size_t nBytes){
     void* p = malloc(nBytes);
     printf("\t...malloc(%d): %x\n", nBytes, p);
     return p;
